# openapi.yaml
...
openapi: 3.1.0
info:
  version: 1.0.0
  title: Grocery Manager API
  description: OpenAPI document of Grocery Manager backend routes
servers:
  - url: https://grocery-manager-qsr9.onrender.com
    description: Main production server 

# Paths/Routes ================================================================
paths:
...
  /users/groups:
    get: 
      summary: Get all group names joined by users
      security:
        - bearerAuth: []
      responses:
        '200':
          description: A list of groups the user is a member of
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    group_id:
                      type: integer
                      description: The ID of the group
                    group_name:
                      type: string
                      description: The name of the group
                    description:
                      type: string
                      description: A brief description of the group
        '404':
          description: No groups associated with user
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /groups/:
    get:
      summary: Get all group information
      responses:
        '200':
          description: Group information retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Group'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: Create a group
      description: This endpoint allows creation of a new group by providing the group name and an optional description
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                group_name:
                  type: string
                  description: The name of the group. This field is required.
                description:
                  type: string
                  description: A description of the group. This field is optional.
              required:
                - group_name
      responses:
        '201': 
          description: Group created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/ResourceAlreadyExists'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      summary: Update a group
      description: This endpoint updates an existing group by providing a new group name and an optional description
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Group'
      responses:
        '200': 
          description: Group updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      summary: Delete a group
      description: This endpoint deletes a group by providing a group name
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                group_name:
                  type: string
                  description: Name of the group to be deleted
      responses:
        '204':
          description: Group deleted successfully
        '404':
          $ref: '#/components/responses/NotFoundError'
  /groups/resolve/{group_name}:
    get:
      summary: Get the group ID for a given group name
      parameters:
        - name: group_name
          in: path
          required: true
          schema:
            type: string
      
      responses:
        '200':
          description: Group successfully found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: 
                    type: string
                    example: "Group resolved successfully"
                  group_id:
                    type: int
                    description: Group ID of the given group name
        '404':
          description: No group found with name '{group_name}'
          $ref: '#/components/responses/NotFoundError'
  /groups/{group_id}:

    parameters:
      - name: group_id
        in: path
        required: true
        schema:
          type: integer

    get:
      summary: Get information of the group with specified group_id
      responses:
        '200':
          description: Group information retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '404':
          description: Group not found
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      summary: This endpoint
      responses:
        '200':
          description: Group information updated successfully
        '404':
          description: Group not found
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete: 
      summary: Delete a group based on its ID.
      responses:
        '204':
          description: Group deleted successfully
        '404':
          description: Group not found
        '500':
          $ref: '#/components/responses/InternalServerError'
  /groups/{group_id}/users:

    parameters:
    - name: group_id
      in: path
      required: true
      schema:
        type: integer

    get:
      summary: Get a list of all users within the group
      responses:
        '200':
          description: A list of users successfully found
          content:
            application/json:
              type: array
              items:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          description: No users found within this group
          $ref: '#/components/responses/NotFoundError'
  /groups/{group_id}/users/<user_id>:

    parameters:
    - name: group_id
      in: path
      required: true
      schema:
        type: integer
    - name: user_id
      in: path
      required: true
      schema:
        type: integer

    post:
      summary: Add a user to the group given user ID and group ID in path
      responses:
        '200':
          description: User successfully added to group
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: User successfully added to group
                    example: User added to group
        '404':
          description: Group or user does not exist
          $ref: '#/components/responses/NotFoundError'
        '409':
          description: User is already in the group
          $ref: '#/components/responses/ResourceAlreadyExists'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    delete:
      summary: Delete a user from a group
      responses:
        '200': 
          description: User successfully removed from group
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: User removed from group
        '404':
          description: User of group does not exist
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # Receipt routes ============================================================
...
components:
...
  schemas:
    Group:
      type: object
      properties:
        group_id:
          type: integer
        group_name:
          type: string
        description:
          type: string
    User:
      type: object
      properties:
        user_id:
          type: integer
        username:
          type: string
        email:
          type: string
      required:
        - user_id
        - username
        - email
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # Standard responses
  responses:
    BadRequest:
      description: Bad request - malformed request syntax
      content: 
        application/json:
          schema: 
            type: object
            properties:
              error:
                type: string
                example: 'Resource identified cannot be found in request'
    NotFoundError:
      description: Not found - the requested resource does not exist
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: 'No users with this identified is found'
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: 'Internal Server Error'
              message:
                type: string
                example: 'A backend error has occured.'
    ResourceAlreadyExists:
      description: Conflict - Resource already exists
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: 'Resource already exists'
              message:
                type: string
                example: 'The resource you are trying to create or change to already exist'